Renamed 'utill' module to 'util' to amend previous typo. Also renamed in all references to ensure no bugs and tests passed.
